requirements.txt
discord.py
matplotlib
shapely
beautifulsoup4
python-dotenv
aiofiles